refactor(pendulum): Put both classes in the same file
