+++ sqlalchemy_utils/types/enriched_datetime/__init__.py
@@ -1,4 +1,3 @@
 # Module for enriched date, datetime type
 from .arrow_datetime import ArrowDateTime  # noqa
-from .pendulum_date import PendulumDate  # noqa
-from .pendulum_datetime import PendulumDateTime  # noqa
+from .pendulum import PendulumDate, PendulumDateTime  # noqa
